cleanup: Extract types into separate file :wink:
